Move custom input components to "form" folder
