+++ src/lib/components/app/ODHForm.svelte
@@ -1,7 +1,7 @@
 <script lang="ts">
   import { Accordion, AccordionItem, Button, Col, FormGroup, Input, Row } from 'sveltestrap';
   import { createEventDispatcher } from 'svelte';
-  import NIOAutoCoderInput from '$lib/components/app/NIOAutoCoderInput.svelte';
+  import NIOAutoCoderInput from '$lib/components/form/NIOAutoCoderInput.svelte';
   import type { IOResponse, ResourceRetrieveEvent } from '$lib/utils/types';
 
   const resourceDispatch = createEventDispatcher<{ 'update-resources': ResourceRetrieveEvent }>();
